refactor(ui): split useInputModes into matrix, analytics, and result hooks
